Removed properties not yet being used from include_properties_list.csv
Removed properties from this list that we are not currently adding to new Wikidata records. There will likely be some properties to add to this list that have not yet appeared in our test set (e.g., we have "father" here but not "mother" and may use "notable works" for publications).